+++ GiganticEmu.Launcher/ViewModels/LoginContainerViewModel.cs
@@ -5,13 +5,16 @@ namespace GiganticEmu.Launcher;
 
 public class LoginContainerViewModel : ReactiveObject
 {
-    public enum Page { Login, Register };
+    public enum Page { Login, Register, ForgotPassword, ResetPassword };
 
     [Reactive]
     public Page CurrentPage { get; set; } = Page.Login;
 
     [Reactive]
     public bool IsLoading { get; set; } = false;
+
+    [Reactive]
+    public string UserName { get; set; } = "";
 
     [Reactive]
     public UserData? User { get; set; } = null;