+++ GiganticEmu.Shared/Backend/UserResetPWRequest.cs
@@ -1,0 +1,6 @@
+namespace GiganticEmu.Shared;
+
+public record UserPWDeleteRequest()
+{
+    public string UserName { get; init; } = default!;
+}