+++ GiganticEmu.Shared/Backend/UserResetPWResponse.cs
@@ -1,0 +1,5 @@
+namespace GiganticEmu.Shared;
+
+public record UserPWDeleteResponse(RequestResult Code) : ResponseBase(Code)
+{
+}